+++ Sources/APIServer/Utilities/DatabaseConfiguration.swift
@@ -202,5 +202,6 @@ func registerMigrations(on app: Application) {
     app.migrations.add(CreateClassAchievements())
     app.migrations.add(AddSubmissionRetestedByUserID())
     app.migrations.add(AddTestSetupLastRetestedManifestHash())
+    app.migrations.add(CreatePreEnrollments())
     app.migrations.add(SessionRecord.migration)
 }